Speed up the analyze phase: float32 buffer, fused cv2 ops, cv2.dft with a safe fallback
analyze() is the actual per-image bottleneck (an order of magnitude slower
than hashing at realistic photo resolutions), dominated by a shared
float64 grayscale buffer that noise_estimate/blockiness_score both make
multiple full-array passes over. Switching that buffer to float32, and
fusing the numpy multi-pass ops (diff+abs, abs+sum) into single cv2 calls
(cv2.absdiff, cv2.norm) roughly doubles analyze phase throughput on
realistic (12-24MP) photos, measured through the real production code path
(ThreadPoolExecutor, cv2.setNumThreads(1)).

effective_resolution's FFT routes through cv2.dft when the array's
dimensions are DFT-efficient (faster, and properly releases the GIL unlike
numpy's FFT) and falls back to np.fft.fft2 otherwise -- cv2.dft measured up
to ~24x *slower* than numpy on dimensions with large prime factors, since
only np.fft guarantees O(n log n) via Bluestein's algorithm at any size.

Verified against tests/Test-image's real photos: ~1e-7 relative drift on
every metric, zero quality_score keeper-choice flips across all 20 real
duplicate pairs there.

+++ tests/test_heic_support.py
@@ -94,13 +94,13 @@ def test_heic_round_trips_through_hash_and_analysis() -> None:
 
         # compare_image_quality.load_gray: must decode via its own PIL
         # fallback and return BGR uint8 (matching cv2's channel convention)
-        # plus a matching-shape grayscale float64.
+        # plus a matching-shape grayscale float32.
         img, ciq_gray = ciq.load_gray(str(p))
         assert img is not None and img.ndim == 3 and img.shape[2] == 3, (
             f"expected a BGR (H, W, 3) array from load_gray, got shape {getattr(img, 'shape', None)}"
         )
         assert ciq_gray.shape == img.shape[:2], "grayscale array must match the color array's (H, W)"
-        assert ciq_gray.dtype == np.float64
+        assert ciq_gray.dtype == np.float32
 
         # analyze() end to end: the full metrics pipeline must not raise and
         # must return sane, finite values keyed off the real HEIC dimensions.
